Move DefineLink to core atom library
