+++ opencog/atoms/core/DefineLink.cc
