+++ opencog/atoms/core/DefineLink.h
